refactor: split tests_rules in multiple files
